# I Love JavaScript, I Hate Its Frameworks
## Reasons for Loving JavaScript

* Small set of keywords providing so much.
* Object-oriented and functional at the same time with a very simple trick; JavaScript function is also object.
* Ideal for scripting, integrating and interfacing with other technologies.
* Several robust runtimes available for various back and front ends.

## Reasons for hating major JavaScript frameworks:

* Being either for front or back end; none for both.
* Setup, development and deployment are complex and complicated.
* Data-oriented; even users are treated as data.
* Focused mostly on graphical UIs, leading to very poor accessibility for those who can't see graphics; even when optimized for screen readers.
* Built with the concept of Take It All or Leave It.

## Alternative

These are more than enough reasons to put effort on having some alternative.

In this initial commit, I added the README file only containing the word Scenarist marked down as heading of level 1.
Scenarist is the name of the alternative, I'm not sure if this should be called a framework or it's just a trick.
Scenarist is the one who writes scenarios.
That's exactly what I want to have as anything can be expressed in scenarios, whether for front or back end.